requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bleach==5.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
jinja-markdown==1.210911
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
Pygments==2.14.0
pymdown-extensions==9.9.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
webencodings==0.5.1
